requirements.txt
a2a-sdk==0.3.10
httpx==0.25.2
python-dotenv==1.0.0
rich==13.7.0
uuid